// app/[locale]/(routes)/crm/lead-contacts/table-data/schema.ts
export interface LeadContact {
  id: string;
  name: string;
  title?: string;
  phone?: string;
  email?: string;
  lead_id: string;
  lead_company?: string;
  lead_region?: string;
  lead_memo?: string;
  lead_industry?: string;
  lead_company_type?: string;
  lead_employee_scale?: string;
  lead_introduction?: string;
} 